not working but need push

+++ app/server/server.js
@@ -9,6 +9,46 @@ io.set('log level', 2);
 io.sockets.on('connection', function(socket) {
     console.log("========================================");
     
+    socket.on('addNewUser', function(usrName, fn){
+        console.log('...adding new user: '+ usrName) + "with id: " + socket.id;
+
+        if (usersDB.nameNotExist(usrName)) {
+            console.log('//userNotExist');
+
+            usersDB.addNewUser(socket.id, usrName);
+            emitInfoAboutNewUser(usrName);
+            emitUsersList();
+            fn(true);
+        } else if(usersDB.socketIdExist(socket.id)) {
+            console.log('exist in db');
+        } else{
+            userNameAlreadyInUse(socket.id, usrName);
+            fn(false);
+        }
+    });
+
+    socket.on('reconnectUser', function(data){
+        console.log('reconnecting user', data.id, data.userName);
+        usersDB.addReconnectedUser(socket.id, data);
+        emitInfoAboutNewUser(data.userName);
+        emitUsersList();
+    });
+
+    socket.on('logoutUser', function (id) {
+        console.log('logoutUser...');
+        var usr = usersDB.findUserById(id);
+        console.log('userToLogout: ', usr);
+        if(usr !== undefined){
+            usersDB.removeUser(id);
+            userLeft(usr.userName);
+            emitUsersList();
+        }
+        usersDB.removeIdFromTmp(id);
+        socket.disconnect();
+    })
+
+
+
     socket.on('set_username', function(userName) {
         console.info('set username:', userName);
         console.info('for id:', socket.id);
@@ -60,14 +100,44 @@ io.sockets.on('connection', function(socket) {
             });
         }
     })
+
     socket.on('disconnect', function() {
         console.log('//userDisconnect')
         var usr = usersDB.findUserById(socket.id);
-        usersDB.removeUser(socket.id);
-
-        userLeft(usr.userName);
+        console.log('urs: ', usr);
+        if(usr !== undefined){
+            usersDB.removeUser(socket.id);
+            userLeft(usr.userName);
+            emitUsersList();
+        }
     })
 })
+
+
+var emitInfoAboutNewUser = function (uName) {
+    console.log('....emitting info about new user');
+    var list = usersDB.getList();
+    list.forEach(function(user){
+        io.sockets.sockets[user.id].emit('newUserJoin', {
+            "userName": uName
+            ,"amount": list.length
+            ,"userColor": user.color
+            ,'id': user.id
+        });
+        //io.sockets.sockets[user.id].emit('newUserJoin', uName);
+    });
+}
+
+var emitUsersList = function () {
+    console.log('....emitting users list')
+    var list = usersDB.getList();
+    if(list.length > 0){
+        list.forEach(function(user){
+            io.sockets.sockets[user.id].emit('usersList', list);
+        });
+    }
+}
+
 
 function userJoined(uName) {
     console.log('//userJoined')
@@ -101,9 +171,7 @@ function userNameAvailable(sId, uName) {
 }
 
 function userNameAlreadyInUse(sId, uName) {
-    setTimeout(function() {
-        io.sockets.sockets[sId].emit('error', {
-            "userNameInUse": true
-        });
-    }, 500);
+    io.sockets.sockets[sId].emit('error', {
+        "userNameInUse": true
+    });
 }
